CRUD ,AUTHENTICATION , AUTHORIZATION , VALIDATION DONE

+++ src/controller/department.contoller.ts
@@ -6,6 +6,8 @@ import { validate } from "class-validator";
 import { plainToInstance } from "class-transformer";
 import { CreateDepartmentDto } from "../dto/department.dto";
 import authorize from "../middleware/authorization";
+import { RequestWithUser } from "../utils/requestwithuser";
+import { Role } from "../utils/enum";
 
 class DepartmentController {
   public router: express.Router;
@@ -15,12 +17,14 @@ class DepartmentController {
 
     this.router.get("/", this.getAllDepartment);
     this.router.get("/:id", this.getDepartmentById);
-    this.router.post("/", this.createDepartment);
-    this.router.delete("/:id", this.deleteDepartment);
+    this.router.post("/", authorize, this.createDepartment);
+    this.router.put("/", authorize, this.updateDepartment);
+    this.router.delete("/:id", authorize, this.deleteDepartment);
   }
 
   public getAllDepartment = async (
-    req: express.Request,
+    req: RequestWithUser,
+
     res: express.Response
   ) => {
     const department = await this.Departmentservice.getAllDepartment();
@@ -28,7 +32,8 @@ class DepartmentController {
   };
 
   public getDepartmentById = async (
-    req: express.Request,
+    req: RequestWithUser,
+
     res: express.Response,
     next: express.NextFunction
   ) => {
@@ -48,12 +53,21 @@ class DepartmentController {
   };
 
   public createDepartment = async (
-    req: express.Request,
+    req: RequestWithUser,
+
     res: express.Response,
     next: express.NextFunction
   ) => {
     const { name } = req.body;
     try {
+      const role = req.role;
+      if (role !== Role.HR) {
+        throw new HttpException(
+          403,
+          "You are not authorized to create employee"
+        );
+      }
+
       const departmentDto = plainToInstance(CreateDepartmentDto, req.body);
       const errors = await validate(departmentDto);
 
@@ -72,25 +86,48 @@ class DepartmentController {
   };
 
   public deleteDepartment = async (
-    req: express.Request,
-    res: express.Response
+    req: RequestWithUser,
+
+    res: express.Response,
+    next: express.NextFunction
   ) => {
-    const department = await this.Departmentservice.DeleteById(
-      Number(req.params.id)
-    );
-    res.status(200).send(department);
+    try {
+      const role = req.role;
+      if (role !== Role.HR) {
+        throw new HttpException(
+          403,
+          "You are not authorized to create employee"
+        );
+      }
+
+      const department = await this.Departmentservice.DeleteById(
+        Number(req.params.id)
+      );
+      res.status(200).send(department);
+    } catch (err) {
+      next(err);
+    }
   };
 
   //update employee
 
   public updateDepartment = async (
-    req: express.Request,
+    req: RequestWithUser,
+
     res: express.Response,
     next: express.NextFunction
   ) => {
     try {
+      const role = req.role;
+      if (role !== Role.HR) {
+        throw new HttpException(
+          403,
+          "You are not authorized to create employee"
+        );
+      }
       const departmentDto = plainToInstance(CreateDepartmentDto, req.body);
       const errors = await validate(departmentDto);
+
       if (errors.length) {
         console.log(JSON.stringify(errors));
         throw new HttpException(400, JSON.stringify(errors));
